No tables or figures were changed. Just added in the text from 'Introduction' and part of my 'Results and Discussion'
It looks like we are going to need to get creative when assembling this final document since formatting changes like centering figure captions below figures seem to work best in word, but figures look terrible when knit to word. That said, I'm going to continue writing the 'Results and Discussion' section and worry about formatting a bit later.